BACKEND/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
